+++ server/migrations/000023_managed_instance_tags.down.sql
@@ -1,0 +1,2 @@
+-- 000023_managed_instance_tags.down.sql
+ALTER TABLE instance_tags DROP COLUMN IF EXISTS managed;